fix(ui): relocate ShortcutBadge to primitives
ShortcutBadge is a leaf <kbd> atom that composes nothing, so it belongs
in primitives/ per catalog authority, not patterns/. Move the file,
repoint the package.json export, SidebarNavRow's relative import, all
external consumer imports, the taxonomy moves.tsv row, and the
component-library.md index (drop it from Patterns, add it to
Primitives, and reword SidebarNavRow's row so it no longer claims a
pattern composes another pattern).
