+++ apps/packages/product-client/src/components/workspace/shell/right-panel/RightPanelHeaderActions.tsx
@@ -1,5 +1,5 @@
 import { IconButton } from "@proliferate/ui/primitives/IconButton";
-import { ShortcutBadge } from "@proliferate/ui/patterns/ShortcutBadge";
+import { ShortcutBadge } from "@proliferate/ui/primitives/ShortcutBadge";
 import { Settings, SplitPanel } from "@proliferate/ui/icons";
 import { SHORTCUTS } from "#product/config/shortcuts/registry";
 import { getShortcutDisplayLabel } from "#product/lib/domain/shortcuts/matching";
