+++ apps/packages/ui/src/primitives/ShortcutBadge.tsx
